Growth School Project Claude/api/requirements.txt
tensorflow==2.13.0
pillow==10.0.0
numpy==1.24.3
google-generativeai==0.3.1
python-dotenv==1.0.0 